refactor(web): split shared @/types into per-feature model folders
Each feature now owns its DTOs under features/{feature}/model/index.ts
instead of a single central src/types/index.ts, so a feature's data
shape lives next to the code that uses it.
